requirements.txt
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
pytest==7.4.0
playwright==1.39.0
psycopg2-binary==2.9.7
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.6.0
pendulum==2.1.2
beautifulsoup4==4.12.2
requests==2.31.0
sqlalchemy==1.4.47
discord.py==2.3.1
python-dotenv==1.0.0
flake8==6.1.0
black==23.7.0
sqlfluff==2.3.0
pydantic==1.10.8
pre-commit==3.3.3
rich==13.4.2
fastapi==0.103.1
uvicorn==0.23.2
Flask-Session==0.4.0
connexion[swagger-ui]==2.14.2
pytest-asyncio